chore: gitkeep 삭제
