+++ src/api/bingsoo/apis.ts
@@ -1,0 +1,2 @@
+import { api } from "../../config/axios";
+import { QueryBingsooRequest, CreateBingsooRequestBody,ChangeFlavorRequestBody } from "./types";
