+++ src/api/bingsoo/types.ts
@@ -1,0 +1,15 @@
+export interface CreateBingsooRequestBody {
+    taste: string;
+    userId: string;
+  }
+
+
+  export interface ChangeFlavorRequestBody {
+    taste: string;
+    userId: string;
+  }
+
+  export interface QueryBingsooRequest{
+    bingsooId: number;
+  }
+
